+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed backup restore not working if one of the backup-parts already existed on the cluster
 - Fixed K8s migrations to use correct generated DB table instances
 - Deleted invalid SpaceHistory entries from K8s.
+- Fixed left-over tiebreaker after evacuation of sinlge DRBD resource
 
 ## [1.26.2] - 2024-02-28